+++ axelor-bank-payment/src/main/resources/i18n/messages.csv
@@ -524,6 +524,7 @@
 "Partner is missing.",,,
 "Password",,,
 "Payment Mode",,,
+"Payment move",,,
 "Payment Reason",,,
 "Payment Status Report",,,
 "Payment Status Report file format",,,
