+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/service/expense/ExpenseService.java
@@ -129,4 +129,6 @@ public interface ExpenseService {
 
   public List<KilometricAllowParam> getListOfKilometricAllowParamVehicleFilter(
       ExpenseLine expenseLine, Expense expense) throws AxelorException;
+
+  public Move createMoveForExpensePayment(Expense expense) throws AxelorException;
 }
