+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/service/expense/ExpenseServiceImpl.java
@@ -24,6 +24,7 @@ import com.axelor.apps.account.db.AnalyticDistributionTemplate;
 import com.axelor.apps.account.db.AnalyticMoveLine;
 import com.axelor.apps.account.db.Invoice;
 import com.axelor.apps.account.db.InvoiceLine;
+import com.axelor.apps.account.db.Journal;
 import com.axelor.apps.account.db.Move;
 import com.axelor.apps.account.db.MoveLine;
 import com.axelor.apps.account.db.PaymentMode;
@@ -36,8 +37,10 @@ import com.axelor.apps.account.service.AccountingSituationService;
 import com.axelor.apps.account.service.AnalyticMoveLineService;
 import com.axelor.apps.account.service.app.AppAccountService;
 import com.axelor.apps.account.service.invoice.generator.InvoiceLineGenerator;
+import com.axelor.apps.account.service.move.MoveCancelService;
 import com.axelor.apps.account.service.move.MoveLineService;
 import com.axelor.apps.account.service.move.MoveService;
+import com.axelor.apps.account.service.payment.PaymentModeService;
 import com.axelor.apps.bankpayment.db.BankOrder;
 import com.axelor.apps.bankpayment.db.repo.BankOrderRepository;
 import com.axelor.apps.bankpayment.service.bankorder.BankOrderService;
@@ -100,6 +103,7 @@ public class ExpenseServiceImpl implements ExpenseService {
   protected AnalyticMoveLineService analyticMoveLineService;
   protected HRConfigService hrConfigService;
   protected TemplateMessageService templateMessageService;
+  protected PaymentModeService paymentModeService;
 
   @Inject
   public ExpenseServiceImpl(
@@ -112,7 +116,8 @@ public class ExpenseServiceImpl implements ExpenseService {
       AccountingSituationService accountingSituationService,
       AnalyticMoveLineService analyticMoveLineService,
       HRConfigService hrConfigService,
-      TemplateMessageService templateMessageService) {
+      TemplateMessageService templateMessageService,
+      PaymentModeService paymentModeService) {
 
     this.moveService = moveService;
     this.expenseRepository = expenseRepository;
@@ -124,6 +129,7 @@ public class ExpenseServiceImpl implements ExpenseService {
     this.analyticMoveLineService = analyticMoveLineService;
     this.hrConfigService = hrConfigService;
     this.templateMessageService = templateMessageService;
+    this.paymentModeService = paymentModeService;
   }
 
   @Override
@@ -537,15 +543,20 @@ public class ExpenseServiceImpl implements ExpenseService {
           Beans.get(BankOrderCreateServiceHr.class).createBankOrder(expense, bankDetails);
       expense.setBankOrder(bankOrder);
       Beans.get(BankOrderRepository.class).save(bankOrder);
-    }
-
-    if (paymentMode.getAutomaticTransmission()) {
       expense.setPaymentStatusSelect(InvoicePaymentRepository.STATUS_PENDING);
     } else {
-      expense.setPaymentStatusSelect(InvoicePaymentRepository.STATUS_VALIDATED);
-      expense.setStatusSelect(ExpenseRepository.STATUS_REIMBURSED);
+      if (accountConfigService
+          .getAccountConfig(expense.getCompany())
+          .getGenerateMoveForInvoicePayment()) {
+        this.createMoveForExpensePayment(expense);
+      }
+      if (paymentMode.getAutomaticTransmission()) {
+        expense.setPaymentStatusSelect(InvoicePaymentRepository.STATUS_PENDING);
+      } else {
+        expense.setPaymentStatusSelect(InvoicePaymentRepository.STATUS_VALIDATED);
+        expense.setStatusSelect(ExpenseRepository.STATUS_REIMBURSED);
+      }
     }
-
     expense.setPaymentAmount(
         expense
             .getInTaxTotal()
@@ -554,13 +565,92 @@ public class ExpenseServiceImpl implements ExpenseService {
             .subtract(expense.getPersonalExpenseAmount()));
   }
 
+  public Move createMoveForExpensePayment(Expense expense) throws AxelorException {
+    Company company = expense.getCompany();
+    PaymentMode paymentMode = expense.getPaymentMode();
+    Partner partner = expense.getUser().getPartner();
+    LocalDate paymentDate = expense.getPaymentDate();
+    BigDecimal paymentAmount = expense.getInTaxTotal();
+    BankDetails companyBankDetails = company.getDefaultBankDetails();
+
+    Account employeeAccount;
+
+    Journal journal =
+        paymentModeService.getPaymentModeJournal(paymentMode, company, companyBankDetails);
+
+    MoveLine expenseMoveLine = this.getExpenseEmployeeMoveLineByLoop(expense);
+
+    if (expenseMoveLine == null) {
+      return null;
+    }
+    employeeAccount = expenseMoveLine.getAccount();
+
+    Move move =
+        moveService
+            .getMoveCreateService()
+            .createMove(
+                journal,
+                company,
+                expense.getMove().getCurrency(),
+                partner,
+                paymentDate,
+                paymentMode,
+                MoveRepository.TECHNICAL_ORIGIN_AUTOMATIC);
+
+    move.addMoveLineListItem(
+        moveLineService.createMoveLine(
+            move,
+            partner,
+            paymentModeService.getPaymentModeAccount(paymentMode, company, companyBankDetails),
+            paymentAmount,
+            false,
+            paymentDate,
+            null,
+            1,
+            expense.getExpenseSeq(),
+            null));
+
+    MoveLine employeeMoveLine =
+        moveLineService.createMoveLine(
+            move,
+            partner,
+            employeeAccount,
+            paymentAmount,
+            true,
+            paymentDate,
+            null,
+            2,
+            expense.getExpenseSeq(),
+            null);
+    employeeMoveLine.setTaxAmount(expense.getTaxTotal());
+
+    move.addMoveLineListItem(employeeMoveLine);
+
+    moveService.getMoveValidateService().validate(move);
+    expense.setPaymentMove(move);
+
+    expenseRepository.save(expense);
+
+    return move;
+  }
+
+  protected MoveLine getExpenseEmployeeMoveLineByLoop(Expense expense) {
+    MoveLine expenseEmployeeMoveLine = null;
+    for (MoveLine moveline : expense.getMove().getMoveLineList()) {
+      if (moveline.getCredit().compareTo(BigDecimal.ZERO) > 0) {
+        expenseEmployeeMoveLine = moveline;
+      }
+    }
+    return expenseEmployeeMoveLine;
+  }
+
   @Override
   public void addPayment(Expense expense) throws AxelorException {
     addPayment(expense, expense.getCompany().getDefaultBankDetails());
   }
 
   @Override
-  @Transactional(rollbackOn = {AxelorException.class, Exception.class})
+  @Transactional(rollbackOn = {Exception.class})
   public void cancelPayment(Expense expense) throws AxelorException {
     BankOrder bankOrder = expense.getBankOrder();
 
@@ -570,9 +660,15 @@ public class ExpenseServiceImpl implements ExpenseService {
         throw new AxelorException(
             TraceBackRepository.CATEGORY_INCONSISTENCY,
             I18n.get(IExceptionMessage.EXPENSE_PAYMENT_CANCEL));
-      } else {
+      } else if (bankOrder.getStatusSelect() != BankOrderRepository.STATUS_CANCELED) {
         Beans.get(BankOrderService.class).cancelBankOrder(bankOrder);
       }
+    }
+
+    Move paymentMove = expense.getPaymentMove();
+    if (paymentMove != null) {
+      expense.setPaymentMove(null);
+      Beans.get(MoveCancelService.class).cancel(paymentMove);
     }
     expense.setPaymentStatusSelect(InvoicePaymentRepository.STATUS_CANCELED);
     expense.setStatusSelect(ExpenseRepository.STATUS_VALIDATED);
