+++ axelor-human-resource/src/main/java/com/axelor/apps/hr/web/expense/ExpenseController.java
@@ -341,7 +341,7 @@ public class ExpenseController {
       expenseServiceProvider.get().addPayment(expense);
       response.setReload(true);
     } catch (Exception e) {
-      TraceBackService.trace(e);
+      TraceBackService.trace(response, e);
       response.setException(e);
     }
   }
@@ -359,7 +359,7 @@ public class ExpenseController {
       expenseServiceProvider.get().cancelPayment(expense);
       response.setReload(true);
     } catch (Exception e) {
-      TraceBackService.trace(e);
+      TraceBackService.trace(response, e);
     }
   }
