+++ axelor-human-resource/src/main/resources/domains/Expense.xml
@@ -43,6 +43,7 @@
   	  <many-to-one name="bankOrder" ref="com.axelor.apps.bankpayment.db.BankOrder" readonly="true" title="Bank Order"/>
   	  <decimal name="paymentAmount" title="Payment amount" readonly="true" />
   	  <integer name="paymentStatusSelect" title="Payment Status" selection="invoice.payment.status.select" readonly="true" default="0"/>
+      <many-to-one name="paymentMove" ref="com.axelor.apps.account.db.Move"/>
 
       <string name="fullName" namecolumn="true" readonly="true">
       <![CDATA[
