+++ axelor-human-resource/src/main/resources/i18n/messages_fr.csv
@@ -514,6 +514,7 @@
 "Payment amount","Montant du paiement",,
 "Payment information","Informations paiement",,
 "Payment mode","Mode de paiement",,
+"Payment move","Ecriture de paiement",,
 "Payroll Leave","Congé",,
 "Payroll Leaves","Congés",,
 "Payroll Preparation","Préparation de paie",,
