+++ axelor-human-resource/src/main/resources/views/Expense.xml
@@ -146,6 +146,7 @@
 					<field name="paymentMode" showIf="paymentMode" readonly="true" colSpan="4"/>
 					<field name="paymentDate" showIf="paymentDate" colSpan="4"/>
 					<field name="bankOrder" hidden="true" showIf="bankOrder" colSpan="4"/>
+					<field name="paymentMove" showIf="paymentMove" readonlyIf="paymentMove" colSpan="4" form-view="move-form" grid-view="move-grid"/>
 				</panel>
 			</panel>
 		</panel-tabs>
